лаба 4/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.10.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3